+++ ecd-20200930/alibabacloud_ecd20200930/models/_describe_zones_request.py
@@ -8,12 +8,14 @@ class DescribeZonesRequest(DaraModel):
     def __init__(
         self,
         region_id: str = None,
+        vpc_access_zone: bool = None,
         zone_type: str = None,
     ):
         # The region ID. You can call the [DescribeRegions](https://help.aliyun.com/document_detail/196646.html) operation to query the most recent region list.
         # 
         # This parameter is required.
         self.region_id = region_id
+        self.vpc_access_zone = vpc_access_zone
         # The type of the zone. Default value: `AvailabilityZone`. This value indicates Alibaba Cloud zones.
         self.zone_type = zone_type
 
@@ -28,6 +30,9 @@ class DescribeZonesRequest(DaraModel):
         if self.region_id is not None:
             result['RegionId'] = self.region_id
 
+        if self.vpc_access_zone is not None:
+            result['VpcAccessZone'] = self.vpc_access_zone
+
         if self.zone_type is not None:
             result['ZoneType'] = self.zone_type
 
@@ -37,6 +42,9 @@ class DescribeZonesRequest(DaraModel):
         m = m or dict()
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
+
+        if m.get('VpcAccessZone') is not None:
+            self.vpc_access_zone = m.get('VpcAccessZone')
 
         if m.get('ZoneType') is not None:
             self.zone_type = m.get('ZoneType')